form n/2 team from a given array of length
where n is even and two people are in one team with equal powers. If they do not have equal powers then one has to solve problems equal to the difference of their power to get the same power then only they can team up. We need to find the total number of problems all people should solve so that we can form n/2 teams.